Deal with the data in pipeline

+++ _2_infoExtract.py
@@ -1,3 +1,4 @@
+import sys
 import time
 import random
 import os
@@ -36,34 +37,46 @@ def getrecord(inputdir,outputdir,userpath):
     solvecnt=0
     starttime=time.time()
     fw=dict()
+    err=set()
     for f in files:
-        for line in gzip.open(inputdir+"/"+f,"rb"):
-            info=line.decode("utf-8").strip().split(",")
-	    day=info[1][0:8]
-	    if day not in fw:
-                fw[day]=open(outputdir+"/"+day+".txt","w")
-                fw[day+"traj"]=open(outputdir+"/"+day+"traj.txt","w")
-            if info[0] in {"1","3","6","7"} and info[4]!="da39a3ee5e6b4b0d3255bfef95601890afd80709" and info[6]!="da39a3ee5e6b4b0d3255bfef95601890afd80709":
-                ua=adduser(info[4])
-                ub=adduser(info[6])
-                fw[day].write("%s,%s,%s-%s,%d,%d\n"%(info[0],info[1][0:17],info[2],info[3],ua,ub))
-            if info[0]=="13" and info[4]!="da39a3ee5e6b4b0d3255bfef95601890afd80709":
-                ua=adduser(info[4])
-                fw[day+"traj"].write("%s,%s-%s,%d\n"%(info[1][0:17],info[2],info[3],ua))
-        solvecnt+=1
-        h,m,s=timeEvaluation(solvecnt,totalcnt,time.time()-starttime)
-        print("getrecord[%s] usercnt: %d, completed: %d/%d -> %.3f%%, remain %02d:%02d:%02d\n"%(inputdir,usercnt,solvecnt,totalcnt,solvecnt*100.0/totalcnt,h,m,s))
+        try:
+            for line in gzip.open(inputdir+"/"+f,"rb"):
+                info=line.decode("utf-8").strip().split(",")
+                day=info[1][0:8]
+                if day not in fw:
+                    fw[day]=open(outputdir+"/"+day+".txt","w")
+                    fw[day+"traj"]=open(outputdir+"/"+day+"traj.txt","w")
+                if info[0] in {"1","3","6","7"} and info[4]!="da39a3ee5e6b4b0d3255bfef95601890afd80709" and info[6]!="da39a3ee5e6b4b0d3255bfef95601890afd80709":
+                    ua=adduser(info[4])
+                    ub=adduser(info[6])
+                    fw[day].write("%s,%s,%s-%s,%d,%d\n"%(info[0],info[1][0:17],info[2],info[3],ua,ub))
+                if info[0]=="13" and info[4]!="da39a3ee5e6b4b0d3255bfef95601890afd80709":
+                    ua=adduser(info[4])
+                    fw[day+"traj"].write("%s,%s-%s,%d\n"%(info[1][0:17],info[2],info[3],ua))
+            solvecnt+=1
+            h,m,s=timeEvaluation(solvecnt,totalcnt,time.time()-starttime)
+            print("getrecord[%s] usercnt: %d, completed: %d/%d -> %.3f%%, remain %02d:%02d:%02d\n"%(inputdir,usercnt,solvecnt,totalcnt,solvecnt*100.0/totalcnt,h,m,s))
+        except:
+            err.add(f)
     for name in fw:fw[name].close()
     fw=open(userpath,"w")
     for u in hashuser:
         fw.write("%s,%d\n"%(u,hashuser[u]))
     fw.close()
-def deletefiles(inputdir):
+    return err
+def deletefiles(inputdir,err):
     for name in os.listdir(inputdir):
-        os.remove(inputdir+"/"+name)
+        if name not in err:os.remove(inputdir+"/"+name)
 if __name__=="__main__":
     hashuser=dict()
-    month=["201412","201501","201502","201503","201504"]
+    month=sys.argv[1:]
     for m in month:
-        getrecord(m,"i"+m,"hash"+m+".txt")
-        deletefiles(m)
+        err=getrecord(m,"i"+m,"hash"+m+".txt")
+        deletefiles(m,err)
+        if len(err)>0:
+            fw=open("err_log.txt","a")
+            t=time.strftime(ISOTIMEFORMAT,time.localtime())
+            for item in err:
+                fw.write("[Error %s]: In _2_infoExtract.py, file %s can't be read\n"%(t,item))
+            fw.close()
+
